backend/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2
alembic
bcrypt
passlib[bcrypt]
python-jose
pydantic
python-dotenv